Keep SlideNav separate from rest of the app.
Previously the Main view was assigned id: "slideNav" by which it
was referenced.  At the same time SlideNav was assigned xtype:
"mainContainer".  Quite a mess.

Now only the Main view has xtype: "mainContainer".  The id is no more
used and no other part of the app no more references SlideNav besides
the Main view which extends it.

Additionally the slideNav action is renamed to toggleNavigation.